Define a variable of modulecmd to avoid 'undeclared var' error
New version (checked with `v4.1.4` on RH8.1) of `modulecmd` always dump a variable with the status if anything was rendered. Since it is dumped without `my`, `eval`'ing it in a strict scope fails. I've addressed that with [PR](https://github.com/cea-hpc/modules/pull/314), but guess this variable can be ignored by MTT.