Run.pm: fix search for MPI Details section
The old code for searching for an MPI Details section was confusing
and overly-complicated.  Additionally, it looked through $ini, which
is a potentially-pruned set of sections (based on the --section CLI
option).

Use a significantly simpler method: we already have the current
section, MPI install section, and MPI get section -- just look for MPI
Details in there.  If there's not one there, skip it.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>